Update NFT pallet and remove soulbound migration

+++ primitives/src/lib.rs
@@ -106,7 +106,7 @@ pub mod types {
 	pub type WeightedDispatchResult = Result<Weight, (Weight, DispatchError)>;
 
 	/// Identifier for a pending issuance of a soulbound token
-	pub type IssuanceId = u64;
+	pub type IssuanceId = u32;
 }
 
 /// Opaque types. These are used by the CLI to instantiate machinery that don't need to know
